refactor(react-ui-list): merge RowList into Listbox; rename Row → Listbox.Item
Collapses the duplicate `RowList` (full-pane picker with ScrollArea Viewport) and
`Listbox` (compact popover variant with `OptionIndicator`) into a single canonical
`Listbox` compound. Both surfaces were doing the same job — single-select listbox
with `aria-selected` + tabster arrow nav — with cosmetic differences only:

- Selection props unified on Radix-style `value` / `onValueChange` (replacing
  RowList's `selectedId` / `onSelectChange`). Single-mode callback narrowed to
  `string` (RowList already filtered undefined).
- `Listbox.Viewport` becomes optional: include it for full-pane pickers (replaces
  the entire RowList API surface); omit it inside popovers/dialogs that own their
  own scroll (replaces the previous bare Listbox shape).
- `Listbox.Item` replaces both the previous `Row` and `Listbox.Option`. Same
  selection wiring, tabIndex=0, role=option, aria-selected, dx-selected.
- `Listbox.ItemLabel` + `Listbox.Indicator` survive for the popover usage (the
  checkmark slot). Indicator now reads its parent Item's selection via a Radix-
  scoped item context.
- `useListboxSelection(id)` replaces `useRowListSelection`.
- New `Popover` story added alongside the existing five — exercises the no-
  Viewport + Indicator combination.

Migrates the seven consumers — six former `RowList` callers (`react-ui-mcp/ToolList`,
`react-ui-introspect/ToolList`, `react-ui-introspect/ToolResults`,
`plugin-meeting/MeetingsList`, `plugin-assistant/TaskList`,
`plugin-integration/IntegrationSettingsArticle`) and the lone former `Listbox` caller
(`plugin-assistant/ChatPrompt/ChatOptions`). All `RowList.*` → `Listbox.*`,
`<Row />` → `<Listbox.Item />`, `selectedId` / `onSelectChange` → `value` /
`onValueChange`; ChatOptions now wraps its options in `Listbox.Content` and uses
the new `ItemLabel` / `Indicator` slot names.

Deletes `packages/ui/react-ui-list/src/components/RowList/` entirely and removes
its `./RowList` re-export from the components barrel.
