+++ packages/plugins/plugin-assistant/src/components/TaskList/TaskList.tsx
@@ -8,7 +8,7 @@ import { Plan } from '@dxos/assistant-toolkit';
 import { useObject } from '@dxos/react-client/echo';
 import { Icon, Tag, type ThemedClassName } from '@dxos/react-ui';
 import { composable, composableProps } from '@dxos/react-ui';
-import { Row, RowList } from '@dxos/react-ui-list';
+import { Listbox } from '@dxos/react-ui-list';
 
 export type TaskListProps = ThemedClassName<{
   plan: Plan.Plan;
@@ -19,11 +19,11 @@ export const TaskList = composable<HTMLDivElement, TaskListProps>(({ plan, ...pr
   const tasks = snapshot?.tasks ?? [];
 
   return (
-    <RowList.Root>
-      <RowList.Viewport {...composableProps(props, { classNames: 'dx-container' })} ref={forwardedRef}>
-        <RowList.Content aria-label='Tasks'>
+    <Listbox.Root>
+      <Listbox.Viewport {...composableProps(props, { classNames: 'dx-container' })} ref={forwardedRef}>
+        <Listbox.Content aria-label='Tasks'>
           {tasks.map((task) => (
-            <Row key={task.id} id={task.id} classNames='py-0'>
+            <Listbox.Item key={task.id} id={task.id} classNames='py-0'>
               <div className='flex items-center gap-2'>
                 <Icon
                   icon={task.status === 'done' ? 'ph--check--regular' : 'ph--circle--regular'}
@@ -36,10 +36,10 @@ export const TaskList = composable<HTMLDivElement, TaskListProps>(({ plan, ...pr
                 <span className='truncate flex-1'>{task.title}</span>
                 {task.status === 'in-progress' && <Tag palette='info'>pending</Tag>}
               </div>
-            </Row>
+            </Listbox.Item>
           ))}
-        </RowList.Content>
-      </RowList.Viewport>
-    </RowList.Root>
+        </Listbox.Content>
+      </Listbox.Viewport>
+    </Listbox.Root>
   );
 });
