--- packages/ui/react-ui-list/src/components/RowList/index.ts
@@ -1,6 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export { RowList, Row, createRowListScope, useRowListSelection } from './RowList';
-export type { RowListRootProps, RowListViewportProps, RowListContentProps, RowProps } from './RowList';
